layout - move restore logic into layout itself (#99972)
* layout - move restore logic into layout itself

* some code :lipstick:
